Separated logic and model
