+++ src/main/java/dev/rednas/moodle/parser/TrueFalseParser.java
@@ -1,29 +1,26 @@
-package dev.rednas.moodle.question.type.truefalse;
+package dev.rednas.moodle.parser;
 
-import dev.rednas.moodle.question.Question;
-import dev.rednas.moodle.question.QuestionType;
-import dev.rednas.moodle.question.type.common.input.selection.RadioButton;
-import lombok.Getter;
+import dev.rednas.moodle.question.common.input.selection.RadioButton;
+import dev.rednas.moodle.question.truefalse.Answer;
+import dev.rednas.moodle.question.truefalse.AnswerBlock;
+import dev.rednas.moodle.question.truefalse.TrueFalseQuestion;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
-@Getter
-public class TrueFalseQuestion extends Question {
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class TrueFalseParser {
 
-    private String questionText;
-    private final AnswerBlock answerBlock = new AnswerBlock();
-
-    private TrueFalseQuestion() {
-        super(QuestionType.TRUEFALSE);
-    }
-
-    public static TrueFalseQuestion createInstance(Element contentElement) {
+    public static TrueFalseQuestion parse(Element contentElement) {
         Elements formulationElement = contentElement.select("div.content > div.formulation");
 
         TrueFalseQuestion question = new TrueFalseQuestion();
-        question.questionText = formulationElement.select("div.qtext").text();
+        question.setQuestionText(formulationElement.select("div.qtext").text());
 
-        question.answerBlock.setPrompt(formulationElement.select("div.prompt").text());
+        AnswerBlock answerBlock = new AnswerBlock();
+        question.setAnswerBlock(answerBlock);
+        answerBlock.setPrompt(formulationElement.select("div.prompt").text());
 
         Elements answerElement = formulationElement.select("div.answer");
         RadioButton radioButton1 = parserRadiobutton(answerElement.select(".r0").first());
@@ -32,7 +29,7 @@ public class TrueFalseQuestion extends Question {
         Answer answer = new Answer();
         answer.setRadioButtonTrue(radioButton1);
         answer.setRadioButtonFalse(radioButton2);
-        question.answerBlock.setAnswer(answer);
+        answerBlock.setAnswer(answer);
 
         return question;
     }
