+++ src/main/java/dev/rednas/moodle/question/Question.java
@@ -4,6 +4,5 @@ import lombok.Data;
 
 @Data
 public abstract class Question {
-    private final QuestionType type;
     private QuestionInfo info;
 }
