+++ src/main/java/dev/rednas/moodle/question/common/input/BaseInput.java
@@ -1,0 +1,8 @@
+package dev.rednas.moodle.question.common.input;
+
+import lombok.Data;
+
+@Data
+public abstract class BaseInput {
+    private String id;
+}
