+++ src/main/java/dev/rednas/moodle/question/common/input/dropdown/Option.java
@@ -1,0 +1,10 @@
+package dev.rednas.moodle.question.common.input.dropdown;
+
+import lombok.Data;
+
+@Data
+public class Option {
+    private String value;
+    private String text;
+    private boolean selected;
+}
