+++ src/main/java/dev/rednas/moodle/question/common/input/selection/Checkbox.java
@@ -1,0 +1,4 @@
+package dev.rednas.moodle.question.common.input.selection;
+
+public class Checkbox extends BaseSelectionControl {
+}
