+++ src/main/java/dev/rednas/moodle/question/common/input/selection/RadioButton.java
@@ -1,0 +1,4 @@
+package dev.rednas.moodle.question.common.input.selection;
+
+public class RadioButton extends BaseSelectionControl {
+}
