+++ src/main/java/dev/rednas/moodle/question/match/MatchQuestion.java
@@ -1,0 +1,16 @@
+package dev.rednas.moodle.question.match;
+
+import dev.rednas.moodle.question.Question;
+import dev.rednas.moodle.question.common.input.dropdown.Dropdown;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class MatchQuestion extends Question {
+    private String questionText;
+    private List<Dropdown> dropdowns = new ArrayList<>();
+}
