+++ src/main/java/dev/rednas/moodle/question/truefalse/TrueFalseQuestion.java
@@ -1,0 +1,12 @@
+package dev.rednas.moodle.question.truefalse;
+
+import dev.rednas.moodle.question.Question;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class TrueFalseQuestion extends Question {
+    private String questionText;
+    private AnswerBlock answerBlock;
+}
