--- src/main/java/dev/rednas/moodle/question/type/common/input/BaseInput.java
@@ -1,9 +1,0 @@
-package dev.rednas.moodle.question.type.common.input;
-
-import lombok.Data;
-
-@Data
-public abstract class BaseInput {
-    private String id;
-    private final InputType type;
-}
