--- src/main/java/dev/rednas/moodle/question/type/common/input/InputType.java
@@ -1,8 +1,0 @@
-package dev.rednas.moodle.question.type.common.input;
-
-public enum InputType {
-    RADIO,
-    CHECKBOX,
-    TEXT,
-    DROPDOWN
-}
