--- src/main/java/dev/rednas/moodle/question/type/common/input/selection/BaseSelectionControl.java
@@ -1,18 +1,0 @@
-package dev.rednas.moodle.question.type.common.input.selection;
-
-import dev.rednas.moodle.question.type.common.input.BaseInput;
-import dev.rednas.moodle.question.type.common.input.InputType;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public abstract class BaseSelectionControl extends BaseInput {
-    private String label;
-    private boolean selected;
-    private Boolean correct;
-
-    public BaseSelectionControl(InputType type) {
-        super(type);
-    }
-}
