--- src/main/java/dev/rednas/moodle/question/type/common/input/selection/Checkbox.java
@@ -1,9 +1,0 @@
-package dev.rednas.moodle.question.type.common.input.selection;
-
-import dev.rednas.moodle.question.type.common.input.InputType;
-
-public class Checkbox extends BaseSelectionControl {
-    public Checkbox() {
-        super(InputType.CHECKBOX);
-    }
-}
