+++ src/test/java/dev/rednas/moodle/question/truefalse/TrueFalseQuestionTest.java
@@ -1,12 +1,10 @@
-package dev.rednas.moodle.question.type.truefalse;
+package dev.rednas.moodle.question.truefalse;
 
 import dev.rednas.moodle.util.AssertionUtils;
 import dev.rednas.moodle.util.TestUtils;
 import dev.rednas.moodle.parser.QuizParser;
 import dev.rednas.moodle.question.Question;
-import dev.rednas.moodle.question.QuestionType;
-import dev.rednas.moodle.question.type.common.input.InputType;
-import dev.rednas.moodle.question.type.common.input.selection.RadioButton;
+import dev.rednas.moodle.question.common.input.selection.RadioButton;
 import dev.rednas.moodle.quiz.Quiz;
 import org.junit.jupiter.api.Test;
 
@@ -21,7 +19,7 @@ class TrueFalseQuestionTest {
         assertEquals(1, quiz.getQuestions().size());
 
         Question question = quiz.getQuestions().get(0);
-        AssertionUtils.assertQuestionInfo(question, null, "1.00", 1L, "notyetanswered", QuestionType.TRUEFALSE);
+        AssertionUtils.assertQuestionInfo(question, null, "1.00", 1L, "notyetanswered");
 
         TrueFalseQuestion truefalseQuestion = (TrueFalseQuestion) question;
         assertEquals("5 + 5 = 10", truefalseQuestion.getQuestionText());
@@ -29,14 +27,12 @@ class TrueFalseQuestionTest {
 
         RadioButton radioButtonTrue = truefalseQuestion.getAnswerBlock().getAnswer().getRadioButtonTrue();
         assertEquals("q3:1_answertrue", radioButtonTrue.getId());
-        assertEquals(InputType.RADIO, radioButtonTrue.getType());
         assertEquals("True", radioButtonTrue.getLabel());
         assertFalse(radioButtonTrue.isSelected());
         assertNull(radioButtonTrue.getCorrect());
 
         RadioButton radioButtonFalse = truefalseQuestion.getAnswerBlock().getAnswer().getRadioButtonFalse();
         assertEquals("q3:1_answerfalse", radioButtonFalse.getId());
-        assertEquals(InputType.RADIO, radioButtonFalse.getType());
         assertEquals("False", radioButtonFalse.getLabel());
         assertFalse(radioButtonFalse.isSelected());
         assertNull(radioButtonFalse.getCorrect());
@@ -49,7 +45,7 @@ class TrueFalseQuestionTest {
         assertEquals(1, quiz.getQuestions().size());
 
         Question question = quiz.getQuestions().get(0);
-        AssertionUtils.assertQuestionInfo(question, "0.00", "1.00", 1L, "incorrect", QuestionType.TRUEFALSE);
+        AssertionUtils.assertQuestionInfo(question, "0.00", "1.00", 1L, "incorrect");
 
         TrueFalseQuestion truefalseQuestion = (TrueFalseQuestion) question;
         assertEquals("5 + 5 = 10", truefalseQuestion.getQuestionText());
@@ -57,14 +53,12 @@ class TrueFalseQuestionTest {
 
         RadioButton radioButtonTrue = truefalseQuestion.getAnswerBlock().getAnswer().getRadioButtonTrue();
         assertEquals("q3:1_answertrue", radioButtonTrue.getId());
-        assertEquals(InputType.RADIO, radioButtonTrue.getType());
         assertEquals("True", radioButtonTrue.getLabel());
         assertFalse(radioButtonTrue.isSelected());
         assertNull(radioButtonTrue.getCorrect());
 
         RadioButton radioButtonFalse = truefalseQuestion.getAnswerBlock().getAnswer().getRadioButtonFalse();
         assertEquals("q3:1_answerfalse", radioButtonFalse.getId());
-        assertEquals(InputType.RADIO, radioButtonFalse.getType());
         assertEquals("False", radioButtonFalse.getLabel());
         assertTrue(radioButtonFalse.isSelected());
         assertFalse(radioButtonFalse.getCorrect());
@@ -77,7 +71,7 @@ class TrueFalseQuestionTest {
         assertEquals(1, quiz.getQuestions().size());
 
         Question question = quiz.getQuestions().get(0);
-        AssertionUtils.assertQuestionInfo(question, "1.00", "1.00", 1L, "correct", QuestionType.TRUEFALSE);
+        AssertionUtils.assertQuestionInfo(question, "1.00", "1.00", 1L, "correct");
 
         TrueFalseQuestion truefalseQuestion = (TrueFalseQuestion) question;
         assertEquals("5 + 5 = 10", truefalseQuestion.getQuestionText());
@@ -85,14 +79,12 @@ class TrueFalseQuestionTest {
 
         RadioButton radioButtonTrue = truefalseQuestion.getAnswerBlock().getAnswer().getRadioButtonTrue();
         assertEquals("q4:1_answertrue", radioButtonTrue.getId());
-        assertEquals(InputType.RADIO, radioButtonTrue.getType());
         assertEquals("True", radioButtonTrue.getLabel());
         assertTrue(radioButtonTrue.isSelected());
         assertEquals(true, radioButtonTrue.getCorrect());
 
         RadioButton radioButtonFalse = truefalseQuestion.getAnswerBlock().getAnswer().getRadioButtonFalse();
         assertEquals("q4:1_answerfalse", radioButtonFalse.getId());
-        assertEquals(InputType.RADIO, radioButtonFalse.getType());
         assertEquals("False", radioButtonFalse.getLabel());
         assertFalse(radioButtonFalse.isSelected());
         assertNull(radioButtonFalse.getCorrect());
